requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
twilio==9.3.5
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.12
psycopg2-binary==2.9.9
